# Adding sync.Pool to Decompress middleware