@pafuent 2020-11-27
@arun0009 Could you please review this PR?
Sadly I didn't had the time to review the PR in which you added the Decompress middleware, and I saw it when it was merged.
Besides the use of a sync.Pool (which I bring from the Compress middleware), I think that the gzip.Reader could be Close without the defer and that the http.request.Body that is replaced at the end, should be closed. When the Body is not changed, is the Server the one responsible for closing it.