@arun0009 2020-11-27
> Codecov is failing because I'm no testing some error code branches, that sadly I don't know how to test. If someone have any suggestion, I'll be glad to implement it.
> 
> Here are the benchmarks comparison for this change. In this case I did 10 iterations instead of the 5 performed on our [GitHub action](https://github.com/labstack/echo/pull/1699/checks?check_run_id=1461852687)
> 
> ```
> name          old time/op    new time/op    delta
> Decompress-2    2.21µs ± 6%    2.08µs ± 6%   -6.01%  (p=0.003 n=10+10)
> 
> name          old alloc/op   new alloc/op   delta
> Decompress-2    5.49kB ± 0%    4.91kB ± 0%  -10.60%  (p=0.000 n=10+9)
> 
> name          old allocs/op  new allocs/op  delta
> Decompress-2      12.0 ± 0%      11.0 ± 0%   -8.33%  (p=0.000 n=10+10)
> ```

One way of testing InternalServerError or pool error handling is by using interfaces. Here's an example that would cover 91.9% of code base https://gist.github.com/arun0009/f8a19659cf956b82e3abadc2211e3be0