@lammel 2020-12-07
Looks like you also added the missing test for the error path.
Are we ready to merge @pafuent ?